+++ database/migrations/2025_05_23_130141_create_payments_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('payments', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('order_id');
-            $table->string('payment_method');
-            $table->decimal('amount', 10, 2);
-            $table->string('status')->default('pending');
-            $table->timestamps();
+        if (!Schema::hasTable('payments')) {
+            Schema::create('payments', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('order_id');
+                $table->string('payment_method');
+                $table->decimal('amount', 10, 2);
+                $table->string('status')->default('pending');
+                $table->timestamps();
 
-            $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
-        });
+                $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
+            });
+        }
     }
 
     /**
